Extract superclass from EventQueueClassic and put it into EventQueue.h
